Actually apply GL states in dependency order.
'Tis quite intredasting how this could have gone unnoticed for so long. I
can only assume that it is because I have not used any rdeps until now, and
that the implicit dependency that is caused by slot creation has meant that
slot IDs have been assigned in proper order of the same dependencies.